requirements.txt
streamlit
pytesseract
pillow
pandas
xlsxwriter
fpdf